+++ Wrappers/Python/cil/io/NikonDataReader.py
@@ -17,7 +17,7 @@
 from cil.framework import AcquisitionData, AcquisitionGeometry
 from cil.io.TIFF import TIFFStackReader
 import warnings
-import numpy
+import numpy as np
 import os
     
         
@@ -31,7 +31,7 @@ class NikonDataReader(object):
         ----------
 
             
-        file_name: str with full path to .xtexct file
+        file_name: str with full path to .xtekct file
             
         roi: dictionary with roi to load 
                 {'angle': (start, end, step), 
@@ -64,7 +64,7 @@ class NikonDataReader(object):
         Output
         ------
         
-        Acquisition data with corresponding geomrtry, arranged as ['angle', horizontal'] 
+        Acquisition data with corresponding geometry, arranged as ['angle', horizontal'] 
         if a single slice is loaded and ['vertical, 'angle', horizontal'] 
         if more than 1 slices are loaded.
                     
@@ -76,10 +76,6 @@ class NikonDataReader(object):
         self.mode = kwargs.get('mode', 'bin')
         self.fliplr = kwargs.get('fliplr', False)
         
-        if 'normalize' in kwargs.keys():
-            self.normalise = kwargs.get('normalize', True)
-            warnings.warn("'normalize' has now been deprecated. Please use 'normalise' instead.")
-
         if self.file_name is not None:
             self.set_up(file_name = self.file_name,
                         roi = self.roi,
@@ -92,8 +88,7 @@ class NikonDataReader(object):
                roi = {'angle': -1, 'horizontal': -1, 'vertical': -1},
                normalise = True,
                mode = 'bin',
-               fliplr = False,
-               **kwargs):
+               fliplr = False):
         
         self.file_name = file_name
         self.roi = roi
@@ -101,14 +96,10 @@ class NikonDataReader(object):
         self.mode = mode
         self.fliplr = fliplr
         
-        if 'normalize' in kwargs.keys():
-            self.normalise = kwargs.get('normalize', True)
-            warnings.warn("'normalize' has now been deprecated. Please use 'normalise' instead.")
-
         if self.file_name == None:
-            raise Exception('Path to xtek file is required.')
+            raise Exception('Path to xtekct file is required.')
         
-        # check if xtek file exists
+        # check if xtekct file exists
         if not(os.path.isfile(self.file_name)):
             raise Exception('File\n {}\n does not exist.'.format(self.file_name))
         
@@ -118,7 +109,7 @@ class NikonDataReader(object):
         # check labels     
         for key in self.roi.keys():
             if key not in ['angle', 'horizontal', 'vertical']:
-                raise Exception("Wrong label. One of ollowing is expected: angle, horizontal, vertical")
+                raise Exception("Wrong label. One of the following is expected: angle, horizontal, vertical")
         
         roi = self.roi.copy()
         
@@ -131,8 +122,8 @@ class NikonDataReader(object):
         if 'vertical' not in roi.keys():
             roi['vertical'] = -1
                 
-        # parse xtek file
-        with open(self.file_name, 'r') as f:
+        # parse xtekct file
+        with open(self.file_name, 'r', errors='replace') as f:
             content = f.readlines()    
                 
         content = [x.strip() for x in content]
@@ -140,8 +131,11 @@ class NikonDataReader(object):
         #initialise parameters
         detector_offset_h = 0
         detector_offset_v = 0
-        object_offset_x = 0
-        object_roll_deg = 0
+        object_tilt_deg = 0
+        object_offset_x = None
+        object_roll_deg = None
+        centre_of_rotation_top = 0
+        centre_of_rotation_bottom = 0
 
         for line in content:
             # filename of TIFF files
@@ -183,12 +177,22 @@ class NikonDataReader(object):
             # detector offset y in units  
             elif line.startswith("DetectorOffsetY"):
                 detector_offset_v = float(line.split('=')[1])
+            #new file format rotation axis offset and angle
             # object offset x in units  
             elif line.startswith("ObjectOffsetX"):
                 object_offset_x = float(line.split('=')[1])
-            # object roll in degrees  
+            # object roll in degrees  (around z)
             elif line.startswith("ObjectRoll"):
                 object_roll_deg = float(line.split('=')[1])
+            # object tilt in degrees  (around x)
+            elif line.startswith("ObjectTilt"):
+                object_tilt_deg = float(line.split('=')[1])
+            #old file format rotation axis offset and angle, in mm at the detector
+            elif line.startswith("CentreOfRotationTop"):
+                centre_of_rotation_top = float(line.split('=')[1])
+            elif line.startswith("CentreOfRotationBottom"):
+                centre_of_rotation_bottom = float(line.split('=')[1])
+          
             # directory where data is stored
             elif line.startswith("InputFolderName"):
                 input_folder_name = line.split('=')[1]
@@ -230,8 +234,9 @@ class NikonDataReader(object):
             pixel_size_v = pixel_size_v_0 * self._roi_par[1][2]
             pixel_size_h = pixel_size_h_0 * self._roi_par[2][2]
         else: # slice
-            pixel_num_v = numpy.int(numpy.ceil((self._roi_par[1][1] - self._roi_par[1][0]) / self._roi_par[1][2]))
-            pixel_num_h = numpy.int(numpy.ceil((self._roi_par[2][1] - self._roi_par[2][0]) / self._roi_par[2][2]))
+            pixel_num_v = int(np.ceil((self._roi_par[1][1] - self._roi_par[1][0]) / self._roi_par[1][2]))
+            pixel_num_h = int(np.ceil((self._roi_par[2][1] - self._roi_par[2][0]) / self._roi_par[2][2]))
+
             pixel_size_v = pixel_size_v_0
             pixel_size_h = pixel_size_h_0
         
@@ -245,8 +250,8 @@ class NikonDataReader(object):
         det_end = det_start + pixel_num_v * self._roi_par[1][2]
         det_pos_v = (det_start + det_end) * 0.5 * pixel_size_v_0 + detector_offset_v         
 
-        #angles from xtek.ct ignore *.ang and _ctdata.txt as not correct
-        angles = numpy.asarray( [ angular_step * proj for proj in range(num_projections) ] , dtype=numpy.float32)
+        #angles from xtekct ignore *.ang and _ctdata.txt as not correct
+        angles = np.asarray( [ angular_step * proj for proj in range(num_projections) ] , dtype=np.float32)
         
         if self.mode == 'bin':
             n_elem = (self._roi_par[0][1] - self._roi_par[0][0]) // self._roi_par[0][2]
@@ -258,8 +263,32 @@ class NikonDataReader(object):
         #convert NikonGeometry to CIL geometry
         angles = -angles - initial_angle + 180
 
-        object_roll_deg * numpy.pi /180.
-        rotate_axis_x = numpy.tan(object_roll_deg * numpy.pi /180.)
+        if object_offset_x == None:
+            object_offset_x = (centre_of_rotation_bottom+centre_of_rotation_top)* 0.5 *  source_to_origin /source_to_det
+
+        if object_roll_deg == None:
+            x = (centre_of_rotation_top-centre_of_rotation_bottom)
+            y = pixel_num_v_0 * pixel_size_v_0
+            object_roll = -np.arctan2(x, y)
+        else:
+            object_roll = object_roll_deg * np.pi /180.
+
+        object_tilt = -object_tilt_deg * np.pi /180.
+
+        tilt_matrix = np.eye(3)
+        tilt_matrix[1][1] = tilt_matrix[2][2] = np.cos(object_tilt)
+        tilt_matrix[1][2] = -np.sin(object_tilt)
+        tilt_matrix[2][1] = np.sin(object_tilt)
+
+        roll_matrix = np.eye(3)
+        roll_matrix[0][0] = roll_matrix[2][2] = np.cos(object_roll)
+        roll_matrix[0][2] = np.sin(object_roll)
+        roll_matrix[2][0] = -np.sin(object_roll)
+
+        #order of construction may be reversed, but unlikely to have both in a dataset
+        rot_matrix = np.matmul(tilt_matrix,roll_matrix)
+        rotation_axis_direction = rot_matrix.dot([0,0,1])
+
 
         if self.fliplr:
             origin = 'top-left'
@@ -279,7 +308,7 @@ class NikonDataReader(object):
         else:
             self._ag = AcquisitionGeometry.create_Cone3D(source_position=[0, -source_to_origin, 0],
                                                          rotation_axis_position=[-object_offset_x, 0, 0],
-                                                         rotation_axis_direction=[rotate_axis_x,0,1],
+                                                         rotation_axis_direction=rotation_axis_direction,
                                                          detector_position=[-det_pos_h, source_to_det-source_to_origin, det_pos_v])
             self._ag.set_angles(angles, 
                                 angle_unit='degree')
@@ -327,12 +356,12 @@ class NikonDataReader(object):
         if (self.normalise):
             ad.array[ad.array < 1] = 1
             # cast the data read to float32
-            ad = ad / numpy.float32(self._white_level)
+            ad = ad / np.float32(self._white_level)
             
         
         if self.fliplr:
             dim = ad.get_dimension_axis('horizontal')
-            ad.array = numpy.flip(ad.array, dim)
+            ad.array = np.flip(ad.array, dim)
         
         return ad