+++ Wrappers/Python/test/test_PluginsAstra_FBP.py
@@ -1,33 +1,27 @@
 # -*- coding: utf-8 -*-
-#   This work is part of the Core Imaging Library (CIL) developed by CCPi 
-#   (Collaborative Computational Project in Tomographic Imaging), with 
-#   substantial contributions by UKRI-STFC and University of Manchester.
+#  Copyright 2018 - 2022 United Kingdom Research and Innovation
+#  Copyright 2018 - 2022 The University of Manchester
+#
+#  Licensed under the Apache License, Version 2.0 (the "License");
+#  you may not use this file except in compliance with the License.
+#  You may obtain a copy of the License at
+#
+#      http://www.apache.org/licenses/LICENSE-2.0
+#
+#  Unless required by applicable law or agreed to in writing, software
+#  distributed under the License is distributed on an "AS IS" BASIS,
+#  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+#  See the License for the specific language governing permissions and
+#  limitations under the License.
 
-#   Licensed under the Apache License, Version 2.0 (the "License");
-#   you may not use this file except in compliance with the License.
-#   You may obtain a copy of the License at
-
-#   http://www.apache.org/licenses/LICENSE-2.0
-
-#   Unless required by applicable law or agreed to in writing, software
-#   distributed under the License is distributed on an "AS IS" BASIS,
-#   WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-#   See the License for the specific language governing permissions and
-#   limitations under the License.
-
-
-from utils_projectors import TestCommon_FBP_SIM
 import unittest
-from utils import has_gpu_astra, has_astra
+from utils_projectors import TestCommon_FBP_SIM
+from utils import has_astra, has_nvidia, initialise_tests
+
+initialise_tests()
 
 if has_astra:
-    from cil.plugins.astra import ProjectionOperator
     from cil.plugins.astra import FBP
-
-has_astra_gpu = has_gpu_astra()
-if not has_astra_gpu:
-    print("Unable to run ASTRA GPU tests")
-
 
 def setup_parameters(self):
 
@@ -38,7 +32,7 @@ def setup_parameters(self):
 
 class Test_Cone3D_FBP_GPU(unittest.TestCase, TestCommon_FBP_SIM):
 
-    @unittest.skipUnless(has_astra_gpu, "Requires ASTRA GPU")
+    @unittest.skipUnless(has_astra and has_nvidia, "Requires ASTRA GPU")
     def setUp(self):
         setup_parameters(self)
         self.Cone3D()
@@ -48,7 +42,7 @@ class Test_Cone3D_FBP_GPU(unittest.TestCase, TestCommon_FBP_SIM):
 
 class Test_Cone2D_FBP_GPU(unittest.TestCase, TestCommon_FBP_SIM):
 
-    @unittest.skipUnless(has_astra_gpu, "Requires ASTRA GPU")
+    @unittest.skipUnless(has_astra and has_nvidia, "Requires ASTRA GPU")
     def setUp(self):
         setup_parameters(self)
         self.Cone2D()
@@ -58,7 +52,7 @@ class Test_Cone2D_FBP_GPU(unittest.TestCase, TestCommon_FBP_SIM):
 
 class Test_Parallel3D_FBP_GPU(unittest.TestCase, TestCommon_FBP_SIM):
 
-    @unittest.skipUnless(has_astra_gpu, "Requires ASTRA GPU")
+    @unittest.skipUnless(has_astra and has_nvidia, "Requires ASTRA GPU")
     def setUp(self):
         setup_parameters(self)
         self.Parallel3D()
@@ -68,7 +62,7 @@ class Test_Parallel3D_FBP_GPU(unittest.TestCase, TestCommon_FBP_SIM):
 
 class Test_Parallel2D_FBP_GPU(unittest.TestCase, TestCommon_FBP_SIM):
 
-    @unittest.skipUnless(has_astra_gpu, "Requires ASTRA GPU")
+    @unittest.skipUnless(has_astra and has_nvidia, "Requires ASTRA GPU")
     def setUp(self):
         setup_parameters(self)
         self.Parallel2D()
@@ -78,7 +72,7 @@ class Test_Parallel2D_FBP_GPU(unittest.TestCase, TestCommon_FBP_SIM):
 
 class Test_Parallel2D_FBP_CPU(unittest.TestCase, TestCommon_FBP_SIM):
 
-    @unittest.skipUnless(has_astra, "Requires ASTRA")
+    @unittest.skipUnless(has_astra and has_nvidia, "Requires ASTRA")
     def setUp(self):
         setup_parameters(self)
         self.FBP_args['device'] = 'cpu'